Refactor and add FB Auth
